pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pysuika"
version = "0.1.0"
description = "A Python package for the Suika (watermelon-drop) physics game with an agent API."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pygame>=2.0",
    "pymunk>=6.0",
    "numpy>=1.20",
    "pyyaml>=6.0",
]

[project.scripts]
suika = "pysuika.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pysuika.assets" = ["*.png"]
"pysuika.config" = ["*.yaml"]
